inference_code/requirements.txt
requests==2.27.1
smdebug==1.0.12
pyyaml==5.4.1
tqdm==4.64.0
seaborn==0.11.2
psutil==5.9.1
boto3==1.23.7
Pillow==9.4.0
ultralytics==8.0.58